Update patch set 1
Patch Set 1:

The new change here "Can Not Merge", you need to be familiar with git to fix this. However, instead of fixing this, let's learn how to start from scratch in case it's so screwed up. Here's the step:


- remove your learn_java project:
rm -rf ~/git/learn_java

- git clone your project again, after this step, you should have a clean ~/git/learn_java again

- create a new Change

- git pull your new change

- setup Java project again (carefully this time, no Kotlin, and project name is learn_java)

- add a .gitignore file as I mentioned before with this content:

.idea/**/workspace.xml
.idea/**/tasks.xml
.idea/dictionaries
/out/

# add everything to your git (including the .gitignore)
- git add --all

# commit you change
- git commit --amend

# push for review
- git push HEAD:refs/for/master

The .gitignore file is to ignore file like .idea/workspace.xml, out/.../Timus1000.class as what you saw in this change. Those files should not be checked in your repository because they are produced.

Patch-set: 1