Update patch set 4
Patch Set 4: Verified+1 Code-Review+2

Now, let see how to submit this one. You probably figure out by now that this change cannot be rebased by the Rebase button in the UI.

In your terminal, you should be in the branch for this change:
$ git checkout <your_branch>

Then:
$ git pull --rebase origin master

You will see that there 's some conflict here. If you do:
$ git status
it will tell you a bit more.

So now edit each of these java file. When there is a conflict, it's usually in this form:
<<<
AAA
====
BBB
>>>>

You need to rearrange the code so that it does not have these <<< ==== >>> any more.

Once you have the files that you're satisfied, 
$ git add --all
$ git rebase --continue
Then you'll have a new commit that is rebased. You can push this commit back to Gerrit:
$ git push origin HEAD:refs/for/master

Patch-set: 4
Label: Code-Review=+2
Label: Verified=+1